VIKING/visual_branch/requirements.txt
h5py==2.10.0
pyyaml
nltk==3.5
black
pylint
flake8
mypy
skipthoughts
neptune-client